Abstract away acceptNftBuyOffer
