+++ src/XrplSandbox/NFTokens/nftClient.ts
@@ -1,0 +1,26 @@
+import { NFTokenAcceptOffer, TxResponse } from 'xrpl';
+import { Amount } from 'xrpl/dist/npm/models/common';
+import { StateRefProvider } from '../types';
+
+/**
+ * {@link https://xrpl.org/nftokenacceptoffer.html}
+ */
+export const acceptNftBuyOffer = async (
+  stateRefProvider: StateRefProvider,
+  offerIndex: string,
+  brokerFee?: Amount
+): Promise<TxResponse> => {
+  const { client, wallet } = await stateRefProvider();
+
+  const acceptNftBuyOfferPayload: NFTokenAcceptOffer = {
+    TransactionType: 'NFTokenAcceptOffer',
+    Account: wallet.address,
+    BuyOffer: offerIndex,
+  };
+
+  if (brokerFee) {
+    acceptNftBuyOfferPayload.BrokerFee = brokerFee;
+  }
+
+  return client.submitAndWait(acceptNftBuyOfferPayload, { wallet });
+};
